+++ src/components/SidePanel/SidePanel.stories.mdx
@@ -1,6 +1,7 @@
 import { Meta, Story, ArgsTable, Canvas } from '@storybook/addon-docs/blocks'
 import { boolean, number, text, select } from '@storybook/addon-knobs'
 import SidePanel from './SidePanel'
+import SidePanelApp from './SidePanel.storiesHelpers'
 
 <Meta
   title="Components/Overlay/SidePanel"
@@ -22,7 +23,7 @@ import SidePanel from './SidePanel'
 
 <Canvas>
   <Story name="default">
-    <SidePanel></SidePanel>
+    <SidePanelApp />
   </Story>
 </Canvas>
