+++ src/components/SimpleModal/index.js
@@ -1,0 +1,3 @@
+import SimpleModal from './SimpleModal'
+
+export default SimpleModal
