+++ src/hooks/usePortal/index.js
@@ -1,0 +1,3 @@
+import usePortal from './usePortal'
+
+export default usePortal
