+++ src/utilities/node.js
@@ -244,3 +244,11 @@ export const isNodeWithinViewport = options => {
 
   return position < viewportPosition
 }
+
+export function getClosestFocusableParent(element) {
+  return (
+    element.closest(
+      'a, button, input, textarea, select, details,[tabindex]:not([tabindex="-1"]:not([disabled]))'
+    ) || document.body
+  )
+}
